refactor: 장착 타입 파일명 변경
