+++ src/domain/types/equipment.types.ts
