+++ src/infrastructure/prisma/prisma.service.ts
@@ -3,6 +3,12 @@ import { PrismaClient } from '@prisma/client';
 
 @Injectable()
 export class PrismaService extends PrismaClient implements OnModuleInit {
+    constructor() {
+        super({
+            log: ['info', 'warn', 'error'],
+        });
+    }
+
     async onModuleInit() {
         await this.$connect();
     }
